+++ example/android/app/src/main/kotlin/com/joel/inter_isolate_event_channel/inter_isolate_event_channel_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.joel.rx_event_channel.rx_event_channel_example
+package com.joel.inter_isolate_event_channel.inter_isolate_event_channel_example
 
 import io.flutter.embedding.android.FlutterActivity